columns-ckan2dataverse.json: Add organization
There are datasets from ILRI and IFPRI in this CKAN instance, but
we are only interested in the ILRI ones. I am including them in the
patterns because ckanapi-exporter's params argument doesn't seem to
be working. For reference:

    ckanapi-exporter ... --params "{'fq':'organization:ilri'}"

At least if I have the organization in the CSV I can filter those
datasets out during cleaning.